// services/hub-shaparak/src/enums/bank-code.enum.ts
// enums/bank-code.enum.ts
export enum BankCode {
    // Central Bank of Iran
    CBI = '010',
    
    // Major Iranian Banks
    MELLI = '011',          // Bank Melli Iran
    SEPAH = '015',          // Bank Sepah
    MELLAT = '012',         // Bank Mellat
    TEJARAT = '018',        // Bank Tejarat
    SADERAT = '019',        // Bank Saderat Iran
    KESHAVARZI = '016',     // Bank Keshavarzi (Agricultural Bank)
    MASKAN = '014',         // Bank Maskan (Housing Bank)
    REFAH = '013',          // Bank Refah Kargaran
    SAMAN = '056',          // Saman Bank
    PARSIAN = '054',        // Parsian Bank
    PASARGAD = '057',       // Bank Pasargad
    EGHTESAD_NOVIN = '055', // Eghtesad Novin Bank
    KARAFARIN = '053',      // Karafarin Bank
    SINA = '059',           // Sina Bank
    SARMAYEH = '058',       // Sarmayeh Bank
    AYANDEH = '062',        // Ayandeh Bank
    SHAHR = '061',          // Shahr Bank
    DAY = '063',            // Day Bank
    ANSAR = '063',          // Ansar Bank
    RESALAT = '070',        // Resalat Gharzolhasaneh Bank
    TOSEE_TAAVON = '022',   // Tosee Taavon Bank
    GHAVAMIN = '064',       // Ghavamin Bank
    GARDESHGARI = '064',    // Gardeshgari Bank (Tourism Bank)
    
    // Test Bank (for development purposes)
    TEST_BANK = '999',
  }